PTF-010-Code reorganization: implementing a folder by page and separate components, improving code lisibility
