--- src/components/pages/About.tsx
@@ -1,5 +1,0 @@
-import React from "react";
-
-export const About: React.FC = () => {
-  return <h1>About</h1>;
-};
