--- src/components/pages/index.tsx
@@ -1,5 +1,0 @@
-export * from "./About";
-export * from "./Contact";
-export * from "./Home";
-export * from "./Services";
-export * from "./Experience";
