+++ src/main.tsx
@@ -1,9 +1,9 @@
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
-import './index.css'
+import './pages/index.css'
 import App from './App.tsx'
 import { BrowserRouter } from 'react-router'
-import { website_adress } from './data.tsx'
+import { website_adress } from './data/data.tsx'
 
 createRoot(document.getElementById('root')!).render(
   <StrictMode>
