+++ src/pages/index.tsx
@@ -1,0 +1,4 @@
+export * from "../pages/About/Experience";
+export * from "../pages/Contact/Contact";
+export * from "../pages/Home/Home";
+export * from "../pages/Services/Services";
